--- src/Pila_/ExcepcionColeccionVacia.java
@@ -1,14 +1,0 @@
-package Pilas;
-/**
- *
- * @author Retama
- */
-public class ExcepcionColeccionVacia extends RuntimeException{
-
-    public ExcepcionColeccionVacia() {
-    }
-    public ExcepcionColeccionVacia(String mensaje) {
-        super(mensaje);
-    }
-    
-}
