+++ src/Pilas/ExcepcionColeccionVacia.java
@@ -1,0 +1,14 @@
+package Pilas;
+/**
+ *
+ * @author Retama
+ */
+public class ExcepcionColeccionVacia extends RuntimeException{
+
+    public ExcepcionColeccionVacia() {
+    }
+    public ExcepcionColeccionVacia(String mensaje) {
+        super(mensaje);
+    }
+    
+}
